refactor resolver and move create conversation to resolver
